+++ test/util/bspNet/helpers.ts
@@ -43,7 +43,8 @@ export const getContainerIp = async (containerName: string, verbose = false): Pr
     console.log(e);
   }
   console.log(
-    `Error fetching container IP for ${containerName} after ${(maxRetries * sleepTime) / 1000
+    `Error fetching container IP for ${containerName} after ${
+      (maxRetries * sleepTime) / 1000
     } seconds`
   );
   showContainers();
