+++ test/util/rpc.ts
@@ -34,7 +34,8 @@ export const sendCustomRpc = async (url: string, method: string, params = [], ve
   }
 
   console.log(
-    `Error fetching ${method} / ${params} from ${url} after ${(maxRetries * sleepTime) / 1000
+    `Error fetching ${method} / ${params} from ${url} after ${
+      (maxRetries * sleepTime) / 1000
     } seconds`
   );
   throw `Error sending custom RPC to ${url}`;
